Move location of file.py to folder named "lib"
